requirements.txt
args==0.1.0
asgiref==3.7.2
clint==0.5.1
Django==4.2
djangorestframework==3.14.0
flake8==7.0.0
flake8-broken-line==1.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
isort==5.13.2
mccabe==0.7.0
progress==1.6
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
